fix(daemon): isolate source worker runtime
