+++ docs/event-loop-contract-audit.md
@@ -115,13 +115,13 @@ The classifier follows execution home, not API spelling. A direct accessor callb
 - `db:recall.dampening.metadata` (withReadDbAsync)
 - `db:recall.final-candidates.hydrate` (withReadDbAsync)
 - `db:recall.final-candidates.safety` (withReadDbAsync)
-- `obsidian-source-embeddings.ts:630` (withReadDb)
-- `obsidian-source-embeddings.ts:666` (withWriteTx)
-- `obsidian-source-embeddings.ts:682` (withReadDb)
-- `obsidian-source-embeddings.ts:716` (withWriteTx)
-- `obsidian-source-embeddings.ts:777` (withWriteTx)
-- `obsidian-source-embeddings.ts:823` (withReadDb)
-- `obsidian-source-embeddings.ts:846` (withWriteTx)
+- `obsidian-source-embeddings.ts:454` (withReadDb)
+- `obsidian-source-embeddings.ts:490` (withWriteTx)
+- `obsidian-source-embeddings.ts:506` (withReadDb)
+- `obsidian-source-embeddings.ts:540` (withWriteTx)
+- `obsidian-source-embeddings.ts:601` (withWriteTx)
+- `obsidian-source-embeddings.ts:647` (withReadDb)
+- `obsidian-source-embeddings.ts:670` (withWriteTx)
 - `obsidian-source-graph.ts:723` (withWriteTx)
 - `obsidian-source-graph.ts:733` (withWriteTx)
 - `obsidian-source-graph.ts:817` (withWriteTx)
